Create Choice Test Set
Creates the test set that I will be using for the initial MVP of the BookExplorer. The idea here is to pick the "best" option among the books in that list (where best means "least similar"). From that, I should be able to gradually expand the "choice set" to encompass all books.