Utilize Dewey Decimal System
Discovered there was a public API for translating a book's isbn number into a dewey decimal number. By construction, DDC numbers are supposed to classify information about a book's topic.
This adds several approaches for translating the DDC number into a recommendation. Overall this seems to match better to the "test" ratings I've compiled compared to the tensorflow word vector.